Move DamageResult to its own file
