+++ src/field/damage-number-handler.ts
@@ -1,5 +1,5 @@
 import { TextStyle, addTextObject } from "../ui/text";
-import type { DamageResult } from "./pokemon";
+import type { DamageResult } from "#app/@types/damage-result";
 import type Pokemon from "./pokemon";
 import { HitResult } from "#enums/hit-result";
 import { formatStat, fixedInt } from "#app/utils";
