+++ src/interfaces/attack-move-result.ts
@@ -1,5 +1,5 @@
 import type { BattlerIndex } from "#app/battle";
-import type { DamageResult } from "#app/field/pokemon";
+import type { DamageResult } from "#app/@types/damage-result";
 import type { Moves } from "#enums/moves";
 
 export interface AttackMoveResult {
