+++ src/phases/damage-anim-phase.ts
@@ -1,7 +1,7 @@
 import { globalScene } from "#app/global-scene";
 import type { BattlerIndex } from "#app/battle";
 import { BattleSpec } from "#enums/battle-spec";
-import type { DamageResult } from "#app/field/pokemon";
+import type { DamageResult } from "#app/@types/damage-result";
 import { HitResult } from "#enums/hit-result";
 import { fixedInt } from "#app/utils";
 import { PokemonPhase } from "#app/phases/pokemon-phase";
